Solid firebase doc/field access
By using FutureBuilder to read Document snapshots, the current code can 'randomly' access the documents stored in the restaurants collection.
The UI will have to change and plan is to move this to a second page that is navigated to from main menu.
We might have to update how we access these fields in the future, but this works for basic functionalities.